Refactor pull up methods and variables
