+++ FBReader/src/main/java/org/geometerplus/android/fbreader/FBReader.java
@@ -24,17 +24,12 @@ import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
-import android.content.res.Configuration;
 import android.database.sqlite.SQLiteDatabase;
 import android.net.Uri;
 import android.os.Bundle;
 import android.preference.PreferenceManager;
-import android.support.annotation.NonNull;
 import android.text.TextUtils;
 import android.util.Log;
-import android.view.KeyEvent;
-import android.view.Menu;
-import android.view.MenuItem;
 import android.view.ViewGroup;
 import android.view.WindowManager;
 import android.view.accessibility.AccessibilityManager;
@@ -47,12 +42,10 @@ import org.benetech.android.R;
 import org.geometerplus.android.fbreader.api.ApiListener;
 import org.geometerplus.android.fbreader.api.ApiServerImplementation;
 import org.geometerplus.android.fbreader.api.PluginApi;
-import org.geometerplus.android.fbreader.benetech.AccessibleMainMenuActivity;
 import org.geometerplus.android.fbreader.benetech.FBReaderWithNavigationBar;
 import org.geometerplus.android.fbreader.library.KillerCallback;
 import org.geometerplus.android.fbreader.library.SQLiteBooksDatabase;
 import org.geometerplus.android.fbreader.network.bookshare.BookshareDeveloperKey;
-import org.geometerplus.android.fbreader.network.bookshare.Bookshare_Webservice_Login;
 import org.geometerplus.android.fbreader.network.bookshare.subscription.BooksharePeriodicalDataSource;
 import org.geometerplus.android.fbreader.network.bookshare.subscription.MainPeriodicalDownloadService;
 import org.geometerplus.android.fbreader.network.bookshare.subscription.PeriodicalEntity;
@@ -73,7 +66,6 @@ import org.geometerplus.zlibrary.core.library.ZLibrary;
 import org.geometerplus.zlibrary.text.hyphenation.ZLTextHyphenator;
 import org.geometerplus.zlibrary.text.view.ZLTextView;
 import org.geometerplus.zlibrary.ui.android.library.ZLAndroidActivityforActionBar;
-import org.geometerplus.zlibrary.ui.android.library.ZLAndroidApplication;
 import org.geometerplus.zlibrary.ui.android.library.ZLAndroidLibrary;
 import org.geometerplus.zlibrary.ui.android.util.SortUtil;
 
@@ -82,20 +74,13 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.ArrayList;
-import java.util.LinkedList;
-import java.util.List;
 
 public class FBReader extends ZLAndroidActivityforActionBar {
-	public static final String BOOK_PATH_KEY = "BookPath";
-    public static final String PREFS_USER_MANUAL_VERSION = "bks_userManualVersion";
+	public static final String PREFS_USER_MANUAL_VERSION = "bks_userManualVersion";
     public static final String USER_GUIDE_FILE = "User-Guide.epub";
 	public static final String FONTS_ASSET_FOLDER = "fonts";
 	public static final String MINI_HELP_FILE_NAME = "MiniHelp.en.fb2";
     public static final String LOG_LABEL = "GoRead";
-
-    //Added for the detecting whether the talkback is on
-    private AccessibilityManager accessibilityManager;
-    private boolean initialOpen = true;
 
 	protected final static int REPAINT_CODE = 1;
 	protected final static int CANCEL_CODE = 2;
@@ -105,11 +90,7 @@ public class FBReader extends ZLAndroidActivityforActionBar {
 	private int myFullScreenFlag;
 	//private InputAccess inputAccess = new InputAccess(this, true);
 
-	private static final String PLUGIN_ACTION_PREFIX = "___";
-	private final List<PluginApi.ActionInfo> myPluginActions =
-		new LinkedList<PluginApi.ActionInfo>();
-
-    private BooksharePeriodicalDataSource dataSource;
+	private BooksharePeriodicalDataSource dataSource;
 
     public static final String SUBSCRIBED_PERIODICAL_IDS_KEY = "subscribed_periodical_ids";
     public static final String AUTOMATIC_DOWNLOAD_TYPE_KEY = "download_type";
@@ -139,18 +120,6 @@ public class FBReader extends ZLAndroidActivityforActionBar {
 			}
 		}
 	};
-
-	@Override
-	protected ZLFile fileFromIntent(Intent intent) {
-		String filePath = intent.getStringExtra(BOOK_PATH_KEY);
-		if (filePath == null) {
-			final Uri data = intent.getData();
-			if (data != null) {
-				filePath = data.getPath();
-			}
-		}
-		return filePath != null ? ZLFile.createFileByPath(filePath) : null;
-	}
 
 	@Override
 	public void onCreate(Bundle icicle) {
@@ -270,132 +239,6 @@ public class FBReader extends ZLAndroidActivityforActionBar {
     		}
     		return ids;
     	}
-    
-    /**
-     * This is a workaround solution because the Ice Cream Sandwich and later releases of Android
-     * made it so that the options menu will not open on larger sized screens.
-     * This solution is gross, but fixes the problem with the menu and 
-     * maintains backwards compatibility.
-     * http://stackoverflow.com/questions/9996333/openoptionsmenu-function-not-working-in-ics/17903128#17903128
-     * In the future we should replace this with the options overflow menu.
-     */
-    @Override
-    public void openOptionsMenu() {
-        super.openOptionsMenu();
-        Configuration config = getResources().getConfiguration();
-        if ((config.screenLayout & Configuration.SCREENLAYOUT_SIZE_MASK) > Configuration.SCREENLAYOUT_SIZE_LARGE) {
-            int originalScreenLayout = config.screenLayout;
-            config.screenLayout = Configuration.SCREENLAYOUT_SIZE_LARGE;
-            super.openOptionsMenu();
-            config.screenLayout = originalScreenLayout;
-        } else {
-            super.openOptionsMenu();
-        }
-    }
-
- 	@Override
-	public boolean onPrepareOptionsMenu(Menu menu) {
-		final ZLAndroidLibrary zlibrary = (ZLAndroidLibrary)ZLibrary.Instance();
-		if (!zlibrary.isKindleFire() && !zlibrary.ShowStatusBarOption.getValue()) {
-			getWindow().addFlags(WindowManager.LayoutParams.FLAG_FORCE_NOT_FULLSCREEN);
-		}
-
-		changeLoginState(menu);
-
-		return super.onPrepareOptionsMenu(menu);
-	}
-
-	private void changeLoginState(Menu menu) {
-		MenuItem loginMenuItem = menu.findItem(R.id.menu_item_login_bookshare);
-		MenuItem logoutMenuItem = menu.findItem(R.id.menu_item_logout_bookshare);
-
-		final boolean isLoggedintoBookshare = isLoggedintoBookshare();
-		if(isLoggedintoBookshare) {
-			String title = String.format(getString(R.string.signout_button_title_pattern), getCurrentLoggedUsername());
-			logoutMenuItem.setTitle(title);
-		}
-		loginMenuItem.setVisible(!isLoggedintoBookshare);
-		logoutMenuItem.setVisible(isLoggedintoBookshare);
-	}
-
-	protected boolean isLoggedintoBookshare() {
-		SharedPreferences login_preference = PreferenceManager.getDefaultSharedPreferences(this);
-		String username = login_preference.getString(Bookshare_Webservice_Login.USER, "");
-		String password = login_preference.getString(Bookshare_Webservice_Login.PASSWORD, "");
-		if (username == null || username.isEmpty())
-			return false;
-
-		if (password == null || password.isEmpty())
-			return false;
-
-		return true;
-	}
-
-	protected String getCurrentLoggedUsername() {
-		SharedPreferences login_preference = PreferenceManager.getDefaultSharedPreferences(this);
-		String username = login_preference.getString(Bookshare_Webservice_Login.USER, "");
-		return username;
-	}
-
-	@Override
-	public void onOptionsMenuClosed(Menu menu) {
-		super.onOptionsMenuClosed(menu);
-		final ZLAndroidLibrary zlibrary = (ZLAndroidLibrary)ZLibrary.Instance();
-		if (!zlibrary.isKindleFire() && !zlibrary.ShowStatusBarOption.getValue()) {
-			getWindow().clearFlags(WindowManager.LayoutParams.FLAG_FORCE_NOT_FULLSCREEN);
-		}
-	}
-
-	@Override
-	public boolean onOptionsItemSelected(MenuItem item) {
-		final ZLAndroidLibrary zlibrary = (ZLAndroidLibrary)ZLibrary.Instance();
-		if (!zlibrary.isKindleFire() && !zlibrary.ShowStatusBarOption.getValue()) {
-			getWindow().clearFlags(WindowManager.LayoutParams.FLAG_FORCE_NOT_FULLSCREEN);
-		}
-
-		String action = findActionForMenuItem(item.getItemId());
-		Object[] params = findParamsForMenuItemAction(item.getItemId());
-		if(params == null){
-			ZLApplication.Instance().doAction(action);
-		}
-		else {
-			ZLApplication.Instance().doAction(action, params);
-		}
-
-		return super.onOptionsItemSelected(item);
-	}
-
-	@NonNull
-	private String findActionForMenuItem(int itemId) {
-		if (itemId == R.id.menu_item_settings)
-			return ActionCode.SHOW_PREFERENCES;
-
-		if (itemId == R.id.menu_item_sync_with_bookshare) {
-			if (isLoggedintoBookshare())
-				return ActionCode.SYNC_WITH_BOOKSHARE;
-
-			return ActionCode.BOOKSHARE;
-		}
-
-		if (itemId == R.id.menu_item_help)
-			return ActionCode.SHOW_HELP;
-
-		if (itemId == R.id.menu_item_about_goread)
-			return ActionCode.ABOUT_GOREAD;
-
-		if (itemId == R.id.menu_item_logout_bookshare)
-			return ActionCode.LOGOUT_BOOKSHARE;
-
-		if (itemId == R.id.menu_item_login_bookshare)
-			return ActionCode.BOOKSHARE;
-		return "";
-	}
-
-	private Object[] findParamsForMenuItemAction(int itemId){
-		if (itemId == R.id.menu_item_sync_with_bookshare)
-			return new Object[]{SyncReadingListsWithBookshareAction.SyncType.USER_ACTIVATED};
-		return null;
-	}
 
 	@Override
 	protected void onNewIntent(Intent intent) {
@@ -518,14 +361,6 @@ public class FBReader extends ZLAndroidActivityforActionBar {
 	}
 
 	@Override
-	protected FBReaderApp createApplication(ZLFile file) {
-		if (SQLiteBooksDatabase.Instance() == null) {
-			new SQLiteBooksDatabase(this);
-		}
-		return new FBReaderApp(file != null ? file.getPath() : null);
-	}
-
-	@Override
 	public boolean onSearchRequested() {
 		final FBReaderApp fbreader = (FBReaderApp)FBReaderApp.Instance();
 		final FBReaderApp.PopupPanel popup = fbreader.getActivePopup();
@@ -600,83 +435,6 @@ public class FBReader extends ZLAndroidActivityforActionBar {
         
         if (currentBook != null) {
             setTitle(currentBook.getTitle());
-        }
-    }
-
-	private Menu addSubMenu(Menu menu, String id) {
-		final ZLAndroidApplication application = (ZLAndroidApplication)getApplication();
-		return application.myMainWindow.addSubMenu(menu, id);
-	}
-
-	private void addMenuItem(Menu menu, String actionId, String name) {
-		final ZLAndroidApplication application = (ZLAndroidApplication)getApplication();
-		application.myMainWindow.addMenuItem(menu, actionId, null, name);
-	}
-
-	private void addMenuItem(Menu menu, String actionId, int iconId) {
-		final ZLAndroidApplication application = (ZLAndroidApplication)getApplication();
-		application.myMainWindow.addMenuItem(menu, actionId, iconId, null);
-	}
-
-	private void addMenuItem(Menu menu, String actionId) {
-		final ZLAndroidApplication application = (ZLAndroidApplication)getApplication();
-		application.myMainWindow.addMenuItem(menu, actionId, null, null);
-	}
-
-	private void addMenuItem(Menu menu, int itemId, String actionId) {
-		final ZLAndroidApplication application = (ZLAndroidApplication)getApplication();
-		application.myMainWindow.addMenuItem(menu, itemId, actionId, null, null);
-	}
-
-	private void addMenuItem(Menu menu, int itemId, String actionId, String name) {
-		final ZLAndroidApplication application = (ZLAndroidApplication)getApplication();
-		application.myMainWindow.addMenuItem(menu, itemId, actionId, null, name);
-	}
-
-	private void addMenuItem(Menu menu, String actionId, String name, int iconId) {
-    		final ZLAndroidApplication application = (ZLAndroidApplication)getApplication();
-    		application.myMainWindow.addMenuItem(menu, actionId, iconId, name);
-    	}
-
-	@Override
-	public boolean onCreateOptionsMenu(Menu menu) {
-		super.onCreateOptionsMenu(menu);
-		getMenuInflater().inflate(R.menu.toolbar_overflow_menu, menu);
-
-		synchronized (myPluginActions) {
-			int index = 0;
-			for (PluginApi.ActionInfo info : myPluginActions) {
-				if (info instanceof PluginApi.MenuActionInfo) {
-					addMenuItem(menu, PLUGIN_ACTION_PREFIX + index++, ((PluginApi.MenuActionInfo)info).MenuItemName);
-				}
-			}
-		}
-
-		changeLoginState(menu);
-
-		final ZLAndroidApplication application = (ZLAndroidApplication)getApplication();
-		application.myMainWindow.refreshMenu();
-
-		return true;
-	}
-
-	/*
-     * show accessible full screen menu when accessibility is turned on
-     *
-    */
-    public boolean onKeyDown(int keyCode, KeyEvent event) {
-        if (accessibilityManager.isEnabled()) {
-            if(keyCode == KeyEvent.KEYCODE_MENU){
-                Intent i = new Intent(this, AccessibleMainMenuActivity.class);
-                startActivity(i);
-            }
-        }
-        return super.onKeyDown(keyCode, event);
-    }
-    
-    public void onWindowFocusChanged(boolean hasFocus) {
-        if (hasFocus && accessibilityManager.isEnabled() && initialOpen) {
-            initialOpen = false;
         }
     }
